requirements.txt
blinker==1.6.2
click==8.1.5
Flask==2.3.2
gunicorn==20.1.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.0.33
protobuf==3.20.3
psycopg2-binary==2.9.6
Werkzeug==2.3.6
zipp==3.16.2
